--- a/source/Ecs.CSharp.Benchmark/SystemWithTwoComponents/Myriad.cs
+++ b/source/Ecs.CSharp.Benchmark/SystemWithTwoComponents/Myriad.cs
@@ -21,7 +21,7 @@ namespace Ecs.CSharp.Benchmark
                 t0.Value += t1.Value;
             }
 
-            public void Execute(ReadOnlySpan<Entity> e, Span<Component1> t0, Span<Component2> t1)
+            public void Execute(ChunkHandle chunk, ReadOnlySpan<Entity> e, Span<Component1> t0, Span<Component2> t1)
             {
                 for (int i = 0; i < t0.Length; i++)
                 {